chore: fix /trashcan command name & move it to misc
